inline Unification::Core into TypeSystemCore
