+++ src/typechecker/typecheck.cpp
@@ -34,7 +34,7 @@ struct TypecheckHelper {
 	void new_nested_scope() { tc.m_env.new_nested_scope(); }
 	void end_scope() { tc.m_env.end_scope(); }
 
-	void unify(MonoId i, MonoId j) { core().m_mono_core.ll_unify(i, j); }
+	void unify(MonoId i, MonoId j) { core().ll_unify(i, j); }
 
 	bool is_type(MetaTypeId i) { return meta_type_is(i, Tag::Func) || meta_type_is(i, Tag::Mono); }
 	bool is_term(MetaTypeId i) { return meta_type_is(i, Tag::Term); }
